refactor(runtime): move runtime state definition
